training/requirements.txt
customtkinter
pandas
numpy
mediapipe
opencv-python
Pillow
fastapi
uvicorn